panoramix-server/requirements.txt
apimas-drf==0.3
